Make TLS optional for SMTP server connnection
This adds a Boolean EMAILUSETLS parameter in defaults.py that
controls whether to use the TLS protocol when authenticating
to the SMTP server.  By default, it is true.

Also added an authentication attempt when initializing the quemail
service.  That adds an error to the output log at program
launch if the SMTP parameters don't work.  Otherwise, those errors
are not seen until some email is queued for sending.